Use the upstream babel-preset-es2015 with a custom configuration to disable modules
babel-preset-es2015 can now (since?) be configured and the module magic can be switched
off. This makes our own preset pretty much useless.

Replace all code with a simple module that just returns a reference to the es2015 preset
with the 'modules' parameter set to 'false'. This keeps the functionality in place.

Fixes #1 (As side-effect: the babel-plugin-transform-es2015-duplicate-keys is now implicitly included
via the upstream preset reference)